Create FCFSalgo
This is a C program that calculates the waiting time and turnaround time of a set of processes using the First Come First Serve (FCFS) scheduling algorithm. The program asks the user to enter the number of processes and their burst times, and then prints the process ID, burst time, waiting time, and turnaround time for each process. It also calculates and prints the average waiting time and turnaround time for all the processes1